создание заказа из корзины

+++ src/main/java/com/example/onlineshop/repository/ConditionRepository.java
@@ -3,5 +3,5 @@ package com.example.onlineshop.repository;
 import com.example.onlineshop.entity.order.Condition;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ConditionRepository extends JpaRepository<Condition, Long> {
+public interface ConditionRepository extends JpaRepository<Condition, Integer> {
 }
